+++ src/main/java/deponn/depmount/DistanceHeightData.java
@@ -1,0 +1,14 @@
+package deponn.depmount;
+
+/**
+ * 距離の二乗と高さの関係
+ */
+public class DistanceHeightData {
+    public final double xzDistance;
+    public final double yHeight;
+
+    public DistanceHeightData(double xzDistance, double yHeight) {
+        this.xzDistance = xzDistance;
+        this.yHeight = yHeight;
+    }
+}
